+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add placeholder container if empty nav field
 - Add autofocus to app search field on show
 - Add current locale context to fields in fields renderer component
+- Invalidate opcache for edited php files via finder
 
 ## 2.3.4 (2022-11-25)